+++ backend/canisters/community/impl/src/updates/update_bot.rs
@@ -24,7 +24,8 @@ fn update_bot_impl(args: Args, state: &mut RuntimeState) -> OCResult {
 
     let now = state.env.now();
 
-    if !state.data.update_bot(
+    if !state.data.update_bot_permissions(
+        member.user_id,
         args.bot_id,
         args.granted_permissions.clone(),
         args.granted_autonomous_permissions.clone(),