+++ backend/canisters/user_index/impl/src/updates/register_bot.rs
@@ -69,14 +69,11 @@ fn register_bot_impl(args: Args, state: &mut RuntimeState) -> Response {
             name: args.name.clone(),
             owner: owner_id,
             endpoint: args.endpoint.clone(),
-            description: args.definition.description.clone(),
-            commands: args.definition.commands.clone(),
-            autonomous_config: args.definition.autonomous_config.clone(),
+            definition: args.definition.clone(),
             last_updated: now,
             avatar,
             installations: HashMap::new(),
             registration_status: BotRegistrationStatus::Private(args.permitted_install_location),
-            restricted_locations: args.definition.restricted_locations.clone(),
         }),
     );