+++ backend/canisters/user/CHANGELOG.md
@@ -6,9 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Keep bot event subscriptions in sync ([#8600](https://github.com/open-chat-labs/open-chat/pull/8600))
+
 ### Fixed
 
-- Fix BotLifecycleEvent::Uninstalled location ([8533](https://github.com/open-chat-labs/open-chat/pull/8533))
+- Fix BotLifecycleEvent::Uninstalled location ([#8533](https://github.com/open-chat-labs/open-chat/pull/8533))
 
 ## [[2.0.1877-user](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1877-user)] - 2025-08-20